Improve CP potential-overflow handling by using long long values by default.

+++ src/external/Weight.hpp
@@ -194,7 +194,7 @@ namespace MinisatID {
 namespace MinisatID {
 #define NOARBITPREC
 
-typedef int Weight;
+typedef long long Weight;
 //FAST, NO OVERFLOW SUPPORT
 }
